requirements.txt
flask
pyrebase4
pyzbar
Pillow
qrcode

